+++ src/app/locales/en/utils/add-your-space-action.js
@@ -1,0 +1,12 @@
+export default {
+  title: 'Add one of your spaces',
+  modalHeader: 'Add one of your spaces',
+  modalDescription: 'Choose a space that should become a {{relation}} {{recordType}} "{{recordName}}":',
+  dropdownPlaceholder: 'Choose a space...',
+  modalSubmit: 'Add',
+  loading: 'Loading available spaces... {{progress}}',
+  // recordType
+  harvester: 'harvester',
+  // relation
+  sourceFor: 'source for',
+};