+++ src/tests/integration/utils/space-actions/add-harvester-to-space-action-test.js
@@ -15,34 +15,49 @@ import {
   getModalFooter,
 } from '../../../helpers/modal';
 import { suppressRejections } from '../../../helpers/suppress-rejections';
-import { promiseArray } from 'onedata-gui-common/utils/ember/promise-array';
-import { resolve } from 'rsvp';
 import { selectChoose, clickTrigger } from 'ember-power-select/test-support/helpers';
+import { clearStoreAfterEach } from '../../../helpers/clear-store';
 
 describe(
   'Integration | Utility | space-actions/add-harvester-to-space-action',
   function () {
     const { afterEach } = setupRenderingTest();
 
-    beforeEach(function () {
-      const recordManager = lookupService(this, 'record-manager');
-      const harvesters = [{
-        entityId: 'harvesterId',
+    clearStoreAfterEach(afterEach);
+
+    beforeEach(async function () {
+      const userId = 'user_id';
+      const store = lookupService(this, 'store');
+      const user = await store.createRecord('user', {
+        id: store.userGri(userId),
+        username: 'testuser',
+        fullName: 'Test User',
+      }).save();
+      const sessionService = lookupService(this, 'session');
+      sessionService.set('data', {
+        authenticated: {
+          identity: {
+            user: user.entityId,
+          },
+        },
+      });
+      const harvester = await store.createRecord('harvester', {
         name: 'harvester1',
-        constructor: {
-          modelName: 'harvester',
-        },
-      }];
-      sinon.stub(recordManager, 'getUserRecordList')
-        .withArgs('harvester')
-        .resolves({
-          list: promiseArray(resolve(harvesters)),
-        });
+      }).save();
+      this.set('harvester', harvester);
+      const harvesterList =
+        await store.createRecord('harvesterList', { list: [harvester] }).save();
+      user.set('harvesterList', harvesterList);
+      const space = await store.createRecord('space', {
+        name: 'space1',
+      }).save();
+      this.set('space', space);
+      const spaceList =
+        await store.createRecord('spaceList', { list: [space] }).save();
+      user.set('spaceList', spaceList);
+      await user.save();
       this.set('context', {
-        space: {
-          name: 'space1',
-          entityId: 'spaceId',
-        },
+        relatedRecord: space,
       });
     });
 
@@ -79,7 +94,7 @@ describe(
       expect(getModalHeader().querySelector('h1'))
         .to.have.trimmed.text('Add one of your harvesters');
       expect(getModalBody().querySelector('p')).to.have.trimmed.text(
-        'Choose harvester which should consume metadata from space "space1":'
+        'Choose a harvester that should consume metadata from space "space1":'
       );
       expect(getModalFooter().querySelector('.record-selector-submit'))
         .to.have.trimmed.text('Add');
@@ -115,7 +130,10 @@ describe(
         await click(getModalFooter().querySelector('.record-selector-submit'));
         const actionResult = await actionResultPromise;
         expect(addHarvesterStub).to.be.calledOnce;
-        expect(addHarvesterStub).to.be.calledWith('harvesterId', 'spaceId');
+        expect(addHarvesterStub).to.be.calledWith(
+          this.harvester.entityId,
+          this.space.entityId
+        );
         expect(successNotifySpy).to.be.calledWith(sinon.match.has(
           'string',
           'The harvester has been successfully added to the space.'